Use 1 register per server and register workers for each scheduler
Instead of having global registers, keep 1 per scheduler. Also, workers
are now instantiated on start and registered into the scheduler, which
will introspect it and expose public methods of every worker.

NOTE: No filters applied. Probably the register call should be moved
to the worker class and let the worker register itself in a specific
scheduler.